alert clean up and standardization
