+++ web/client/src/routes/(authenticated)/projects/$iamId/$projectNumber/index.tsx
@@ -22,7 +22,6 @@ import { TooltipLabel } from '@/shared/TooltipLabel.tsx';
 import { tooltipDefinitions } from '@/shared/tooltips.ts';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { createFileRoute, Link } from '@tanstack/react-router';
-import { format } from 'date-fns';
 import {
   ChartBarIcon,
   PresentationChartLineIcon,
@@ -47,54 +46,6 @@ const ProjectNotFound = ({ projectNumber }: { projectNumber: string }) => (
     </section>
   </main>
 );
-
-function parseAwardEndDate(value: string | null) {
-  if (!value) {
-    return null;
-  }
-
-  const dateOnlyMatch = /^(\d{4})-(\d{2})-(\d{2})/.exec(value);
-
-  if (dateOnlyMatch) {
-    const year = Number(dateOnlyMatch[1]);
-    const monthIndex = Number(dateOnlyMatch[2]) - 1;
-    const day = Number(dateOnlyMatch[3]);
-    const date = new Date(year, monthIndex, day);
-    const isSameCalendarDate =
-      date.getFullYear() === year &&
-      date.getMonth() === monthIndex &&
-      date.getDate() === day;
-
-    return Number.isNaN(date.getTime()) || !isSameCalendarDate ? null : date;
-  }
-
-  const parsed = new Date(value);
-
-  return Number.isNaN(parsed.getTime())
-    ? null
-    : new Date(parsed.getFullYear(), parsed.getMonth(), parsed.getDate());
-}
-
-export function getAwardEndedAlertText(
-  awardEndDate: string | null,
-  today = new Date()
-) {
-  const endDate = parseAwardEndDate(awardEndDate);
-
-  if (!endDate) {
-    return null;
-  }
-
-  const todayDateOnly = new Date(
-    today.getFullYear(),
-    today.getMonth(),
-    today.getDate()
-  );
-
-  return endDate < todayDateOnly
-    ? `Award ended on ${format(endDate, 'MM/dd/yyyy')}`
-    : null;
-}
 
 function ProjectContent({
   iamId,
@@ -121,7 +72,6 @@ function ProjectContent({
       ? 'discrepancy'
       : 'balanced'
     : undefined;
-  const awardEndedAlertText = getAwardEndedAlertText(summary.awardEndDate);
 
   return (
     <main className="flex-1 min-w-0">
@@ -146,16 +96,6 @@ function ProjectContent({
         )}
         <div className="mt-6">
           <ProjectAlerts
-            awardEndedAlert={
-              awardEndedAlertText
-                ? {
-                    colorClass: summary.isInternal
-                      ? 'alert-accent'
-                      : 'alert-info',
-                    message: awardEndedAlertText,
-                  }
-                : undefined
-            }
             iamId={iamId}
             reconciliationStatus={reconciliationStatus}
             summary={summary}
